+++ runtime/binding-openapi-asyncapi/src/main/java/io/aklivity/zilla/runtime/binding/openapi/asyncapi/config/OpenapiAsyncapiSpecConfig.java
@@ -16,7 +16,7 @@ package io.aklivity.zilla.runtime.binding.openapi.asyncapi.config;
 
 import java.util.Set;
 
-import io.aklivity.zilla.runtime.binding.asyncapi.config.AsyncapiSpecificationConfig;
+import io.aklivity.zilla.runtime.common.asyncapi.config.AsyncapiSpecificationConfig;
 import io.aklivity.zilla.runtime.common.openapi.config.OpenapiSpecificationConfig;
 
 public class OpenapiAsyncapiSpecConfig